+++ compat/src/compat-shallow-renderer/PreactShallowRenderer.ts
@@ -7,7 +7,7 @@ import type {
 } from 'preact/src/internal';
 import { Component, options as rawOptions, isValidElement } from 'preact';
 
-import { getComponent } from '../preact10-internals.js';
+import { getComponent } from '../../../src/preact10-internals.js';
 import {
   diffComponent,
   commitRoot,
@@ -71,7 +71,7 @@ function installOptionsForShallowRender() {
 
   /**
    * If this component is being shallow rendered, remove all hook effect
-   * callbacks to match how enzyme shallow render's React
+   * callbacks to match how Enzyme shallow render's React
    */
   const prevDiffed = options.diffed;
   options.diffed = vnode => {
@@ -126,8 +126,6 @@ function installShallowComponentHooks() {
   };
 }
 
-installShallowComponentHooks();
-
 /**
  * This class mirrors ReactShallowRenderer for the purpose of shallow rendering
  * Preact components. It implements the same API and passes almost the exact
@@ -168,6 +166,7 @@ export default class PreactShallowRenderer {
   private _rendered: ComponentChild = null;
 
   constructor() {
+    installShallowComponentHooks();
     installOptionsForShallowRender();
     this._reset();
   }
@@ -257,7 +256,7 @@ export default class PreactShallowRenderer {
   ) {
     // React's memo function wraps a component. It doesn't create a new VNode in
     // the virtual tree. To be compatible with React's shallow renderer, we'll
-    // try to mimic this behavior so shallow enzyme tests don't need to be aware
+    // try to mimic this behavior so shallow Enzyme tests don't need to be aware
     // of this subtle difference between Preact and React. Further, how Preact
     // implements Memo today may change in the future so we'll try to avoid
     // exposing that implementation detail to users.