+++ compat/src/index.ts
@@ -1,0 +1,3 @@
+import CompatShallowRenderer from './CompatShallowRenderer';
+
+export { CompatShallowRenderer };