+++ compat/test/compat-shallow-renderer/PreactShallowRendererMemo_test.tsx
@@ -29,12 +29,13 @@ function memo<T extends ComponentType<any>>(component: T): T {
   return realMemo<T>(component as any) as any;
 }
 
-// TODO: React.memo doesn't manifest as a separate component in the virtual
+// React.memo doesn't manifest as a separate component in the virtual
 // tree, so calls to `instance.setState` and other methods are expected to
 // operate on the underlying component beneath memo. However in Preact, since
-// `memo` is it's own component, calls to `instance.setState` operate on the
-// memo component and not the real component underneath it. Might need to fix
-// this for proper shallow rendering to work.
+// `memo` is its own component, calls to `instance.setState` operate on the
+// memo component and not the real component underneath it. These tests assert
+// that the compat shallow properly handles memo components and forwards calls
+// to the actual component under `memo`.
 describe('PreactShallowRendererMemo', () => {
   installVNodeTestHook();